Restructure directories to allow for organized expansion
Properly adds "html" subdirectory for all of the HTML parsing stuff, and
separates it into a makefile category.
