+++ source/util.c
@@ -18,7 +18,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "types.h"
+#include <types.h>
 
 char* tempstr = NULL;
